+++ spec/integrations/factory_doctor_spec.rb
@@ -5,7 +5,7 @@ require "spec_helper"
 describe "FactoryDoctor" do
   context "Minitest Integration" do
     it "works when there are bad examples", :aggregate_failures do
-      output = run_minitest("factory_doctor", env: {"FDOC" => "1"})
+      output = run_minitest("factory_doctor", env: {"FDOC" => "1", "FDOC_THRESHOLD" => "0"})
 
       expect(output).to include("FactoryDoctor report")
       expect(output).to include("Total (potentially) bad examples: 4")
@@ -21,7 +21,7 @@ describe "FactoryDoctor" do
     end
 
     it "print message when no bad examples", :aggregate_failures do
-      output = run_minitest("factory_doctor", env: {"FDOC" => "1", "TESTOPTS" => "--name=test_0005_is_ignored"})
+      output = run_minitest("factory_doctor", env: {"FDOC" => "1", "FDOC_THRESHOLD" => "0", "TESTOPTS" => "--name=test_0005_is_ignored"})
 
       expect(output).to include("FactoryDoctor enabled")
       expect(output).to include('FactoryDoctor says: "Looks good to me!"')
@@ -30,7 +30,7 @@ describe "FactoryDoctor" do
 
   context "RSpec integration" do
     it "works when there are bad examples", :aggregate_failures do
-      output = run_rspec("factory_doctor", env: {"FDOC" => "1"})
+      output = run_rspec("factory_doctor", env: {"FDOC" => "1", "FDOC_THRESHOLD" => "0"})
 
       expect(output).to include("FactoryDoctor report")
       expect(output).to include("Total (potentially) bad examples: 4")
@@ -45,7 +45,7 @@ describe "FactoryDoctor" do
     end
 
     it "print message when no bad examples", :aggregate_failures do
-      output = run_rspec("factory_doctor", env: {"FDOC" => "1", "SPEC_OPTS" => '-e "is ignored"'})
+      output = run_rspec("factory_doctor", env: {"FDOC" => "1", "FDOC_THRESHOLD" => "0", "SPEC_OPTS" => '-e "is ignored"'})
 
       expect(output).to include("FactoryDoctor enabled")
       expect(output).to include('FactoryDoctor says: "Looks good to me!"')
@@ -68,7 +68,7 @@ describe "FactoryDoctor" do
       specify "it works", :aggregate_failures do
         output = run_rspec(
           "factory_doctor_stamp",
-          env: {"FDOC" => "1", "FDOC_STAMP" => "fd_ignore"}
+          env: {"FDOC" => "1", "FDOC_STAMP" => "fd_ignore", "FDOC_THRESHOLD" => "0"}
         )
 
         expect(output).to include("7 examples, 0 failures")
@@ -84,7 +84,7 @@ describe "FactoryDoctor" do
 
         output2 = run_rspec(
           "factory_doctor_stamp",
-          env: {"FDOC" => "1"}
+          env: {"FDOC" => "1", "FDOC_THRESHOLD" => "0"}
         )
 
         expect(output2).to include("7 examples, 0 failures")